Report the frame you resolved in, and refuse the cut that fits no row
Three things the run said that were not true.

**A fixed-width cut that fits no row died in the seam.** `{"index": 40}` or
`{"last": 12}` resolves to an empty window on every row with reason
`out_of_range`, and a non-anchored selection is not flat, so the gather
built its index tensor from a rectangle of empty lists — a float tensor,
and `tensors used as indices must be long`. That is the trap §24.6 fixed in
`_flat` and not at the other site; the dtype is stated there too now. But
the dtype was never the answer: a position of a fixed width names the same
number of tokens on every row, so a row it does not fit is a document that
is wrong about its own prompts. It is refused where it is resolved, naming
the op, the rows and the reason — read and write alike. An *anchored* cut
is the other case and still reports per row, because which rows carry a
word is data.

**A tap in the continuation frame reported a prompt position.** It was
given the dummy window the seam needs, and the record was then filled from
the prompt frame — `rows: (0,)` and, under left padding, whatever
`offsets[-start]` happened to be, which reached the written table as
`"positions": [0]` beside a number read at decode step 2. The prompt frame
has nothing true to say about a tap that acted during the decode, so it now
says nothing, and `_continuation` — which has the frame that can — reports
every continuation-frame tap: the decode step, and the token the model
produced there. A steering write at every step reports every step.

**The write refusal named one cause for three.** "its position's text is
not in those prompts" was fixed text, printed for `alignment_ambiguous`
(the text is there, twice) and `out_of_range` (the text is there and the
cut is too wide) as well. It carries the row's own reason now, and names
the fix the ambiguous case wants.

`Where.spelling()` puts a position in a message the way a document writes
it. `locate.continuation` asks the tokenizer which ids stop a row instead
of being told.

+++ causalab_mini/ops/intervene.py
@@ -248,7 +248,9 @@ def _window(tensor: Any, at: Selection, seq_axis: int) -> Any:
         rows, index = _flat(positions, tensor.device)
         return moved[rows, index]
     rows = torch.arange(tensor.shape[0], device=tensor.device)[:, None]
-    index = torch.as_tensor(positions, device=tensor.device)  # (batch, w)
+    # the dtype is stated for the same reason as in `_flat`: a rectangle of
+    # empty windows is a list of empty lists, and that is a float tensor
+    index = torch.as_tensor(positions, dtype=torch.long, device=tensor.device)  # (batch, w)
     return moved[rows, index]
